engine: extract esm/tours/undo.js
